+++ spec/lib/tasks/annual_fund_impact_metrics_activities_spec.rb
@@ -90,8 +90,8 @@ RSpec.describe "rake activities:annual_fund_impact_metrics", type: :task do
 
     expect(result.count).to be 4
     expect(result[0]).to eq "Partner Organisation name,Activity name,RODA ID,Partner Organisation ID,Status\n"
-    expect(result[1]).to eq "#{aardvark_organisation.name},#{review_activity.title},#{review_activity.roda_identifier},#{review_activity.partner_organisation_identifier},#{review_activity.programme_status}\n"
-    expect(result[2]).to eq "#{aardvark_organisation.name},#{decided_activity.title},#{decided_activity.roda_identifier},#{decided_activity.partner_organisation_identifier},#{decided_activity.programme_status}\n"
-    expect(result[3]).to eq "\"#{beis_organisation.name}\",#{completed_activity.title},#{completed_activity.roda_identifier},#{completed_activity.partner_organisation_identifier},#{completed_activity.programme_status}\n"
+    expect(result[1]).to eq "#{aardvark_organisation.name},#{review_activity.title},#{review_activity.roda_identifier},#{review_activity.partner_organisation_identifier},Review\n"
+    expect(result[2]).to eq "#{aardvark_organisation.name},#{decided_activity.title},#{decided_activity.roda_identifier},#{decided_activity.partner_organisation_identifier},Decided\n"
+    expect(result[3]).to eq "\"#{beis_organisation.name}\",#{completed_activity.title},#{completed_activity.roda_identifier},#{completed_activity.partner_organisation_identifier},Completed\n"
   end
 end